Move collect_asyncgen to utils.asyncgen
