+++ scrapy/utils/asyncgen.py
@@ -1,0 +1,5 @@
+async def collect_asyncgen(result):
+    results = []
+    async for x in result:
+        results.append(x)
+    return results
