+++ scrapy/utils/py36.py
@@ -1,11 +1,11 @@
 import warnings
 
 from scrapy.exceptions import ScrapyDeprecationWarning
-from scrapy.utils.python import collect_asyncgen  # noqa: F401
+from scrapy.utils.asyncgen import collect_asyncgen  # noqa: F401
 
 
 warnings.warn(
-    "Module `scrapy.utils.py36` is deprecated, please import from `scrapy.utils.python` instead.",
+    "Module `scrapy.utils.py36` is deprecated, please import from `scrapy.utils.asyncgen` instead.",
     category=ScrapyDeprecationWarning,
     stacklevel=2,
 )
