+++ scrapy/utils/python.py
@@ -355,10 +355,3 @@ class MutableChain:
     @deprecated("scrapy.utils.python.MutableChain.__next__")
     def next(self):
         return self.__next__()
-
-
-async def collect_asyncgen(result):
-    results = []
-    async for x in result:
-        results.append(x)
-    return results
